test: desc_motor_count unsigned 오버플로 회귀 테스트 추가
Int.MAX_VALUE(2,147,483,647)를 넘는 unsigned 32비트 값(최대 4,294,967,295)이
엔티티/JDBC 바인딩에서 잘리거나 부호 반전되지 않고 왕복되는지 검증한다.
@JdbcTypeCode -> columnDefinition 전환(1e9ed51)이 실제로 지켜지는지 확인하는
회귀 테스트다(작업일지 0113 6차 참고).

- DataReceiveAnalysisRepositoryMotorCountTest 추가
- schema.sql: desc_motor_count를 BIGINT로 확장 — 이 @DataJpaTest가 쓰는 기본
  H2는 "INT UNSIGNED" 문법을 파싱하지 못해(SecuranceApplicationTests와 달리
  MODE=MySQL URL이 아니라 임베디드 DB로 치환됨) 오버플로 값을 담을 폭이
  필요했다. 운영 스키마와 100% 동일하지 않다는 이 파일의 기존 원칙과 일치.

+++ securance-domain/src/test/resources/schema.sql
@@ -85,7 +85,14 @@ CREATE TABLE tb_data_rcv_anal (
     desc_operation06        VARCHAR(20) NOT NULL DEFAULT '',
     desc_operation07        VARCHAR(20) NOT NULL DEFAULT '',
     desc_operation08        VARCHAR(20) NOT NULL DEFAULT '',
-    desc_motor_count        INT         NOT NULL DEFAULT 0,
+    -- 실제 컬럼은 int(10) unsigned(최대 4,294,967,295)이지만, 이 @DataJpaTest가 쓰는 기본(비 MySQL
+    -- 호환 모드) H2는 "INT UNSIGNED" 문법 자체를 파싱하지 못한다(SecuranceApplicationTests처럼
+    -- MODE=MySQL URL을 쓰는 컨텍스트에서만 지원됨 — @DataJpaTest는 커스텀 URL을 무시하고 임베디드
+    -- DB로 치환한다). Int.MAX_VALUE(약 21억)를 넘는 값의 오버플로 회귀 테스트
+    -- (DataReceiveAnalysisRepositoryMotorCountTest 참고)가 의미를 가지려면 컬럼이 그 값을 저장할
+    -- 수는 있어야 하므로, 폭이 더 넓은 BIGINT로 대체한다(운영 스키마와 100% 동일하지 않음 — 이
+    -- 파일 상단 주석대로 원래도 최소 재현일 뿐이다).
+    desc_motor_count        BIGINT      NOT NULL DEFAULT 0,
     desc_master_in_total    BIGINT      NOT NULL DEFAULT 0,
     desc_gate_status01      VARCHAR(50) NOT NULL DEFAULT '',
     desc_gate_status02      VARCHAR(50) NOT NULL DEFAULT '',
